+++ src/application.js
@@ -19,7 +19,8 @@ module.exports = (function () {
         access_token: null,
         access_data: null,
         color: 'default',
-        template: 'tiny'
+        template: 'tiny',
+        host: 'secure.xsolla.com'
     };
 
     App.eventTypes = _.extend({}, PaystationEmbedApp.eventTypes);
@@ -44,6 +45,10 @@ module.exports = (function () {
 
         if (!$(this.config.target_element).length) {
             this.throwError('Target element doesn\'t exist in the DOM');
+        }
+
+        if (_.isEmpty(this.config.host)) {
+            this.throwError('Invalid host');
         }
 
         if (this.config.color !== 'dark' && this.config.color !== 'default') {
@@ -96,7 +101,8 @@ module.exports = (function () {
         }
 
         this.api = new Api(request, {
-            sandbox: options.sandbox
+            sandbox: options.sandbox,
+            host: this.config.host
         });
 
         this.render();
@@ -110,12 +116,19 @@ module.exports = (function () {
     App.prototype.open = function (params) {
         this.checkApp();
 
+        var access_data = {purchase: {tips: undefined}};
+
+        if (params.tips && params.tips.amount && params.tips.currency) {
+            access_data.purchase.tips = params.tips;
+        }
+
         PaystationEmbedApp.init({
             access_token: this.config.access_token,
-            access_data: this.config.access_data,
+            access_data: _.merge(access_data, this.config.access_data),
             sandbox: this.config.sandbox,
             lightbox: this.config.lightbox,
-            childWindow: this.config.childWindow
+            childWindow: this.config.childWindow,
+            host: this.config.host
         });
 
         // Register events (forwarding)
@@ -210,6 +223,7 @@ module.exports = (function () {
                     currency: info.min_currency,
                     hasDifferent: _.uniq(_.pluck(info.drm, 'amount')).length > 1 || _.uniq(_.pluck(info.drm, 'currency')).length > 1
                 },
+                tips: info.tips,
                 name: info.name,
                 description: info.description,
                 systemRequirements: info.system_requirements,